8th commit

+++ routes/blogs.js
@@ -27,18 +27,18 @@ router.get('/nonprivate', authMiddleware,async (req,res) => {
 }
 })
 
-router.post('/like/:id', authMiddleware,async (req,res) => {
+router.put('/like/:id', authMiddleware,async (req,res) => {
     const id = req.params.id
     let blog = ""
     try {
          //* find the BLOG by the id
          //* allow user to like a blog post only once
         let like = await likeModel.find({ blog_entry_id: id, created_by: req.user.id }) // look for a record of like for this blog post for this user
-        console.log(like)
-        if (like.length == 0) {
+        
+        if (like.length === 0) {
            like = await likeModel.create({ blog_entry_id: id, created_by: req.user.id, like: true}) // create a like object to record a like for the post
-           blog = await blogModel.findByIdAndUpdate(id,  { $inc: { likes: 1 } })
-            res.status(202).json(blog)
+           blog = await blogModel.findByIdAndUpdate(id,  { $inc: { likes: 1 }}, {new:true} )
+          res.status(202).json(blog)
         } else {
           blog = await blogModel.findById(id)
             res.status(202).json(blog)
@@ -87,14 +87,17 @@ router.get('/:id', authMiddleware,async (req, res) => {
 router.put('/:id',authMiddleware ,async (req, res) => {
     const id = req.params.id
     const newBlogData = req.body
-    console.log(req.user.id)
-    if (newBlogData.creator_id !== req.user.id) { // only original creator can change blog content
-       return res.status(400).json({msg: 'Not Authorized to Change this Blog'})
-    }
+    
     
     try {
+         const blogtoUpdate = await blogModel.findById(id) 
+         
+         if (blogtoUpdate.creator_id.toString()  !==  req.user.id ) {
+             return res.status(400).json({msg: 'Not Authorized ! '})
+         }
+
          //* find the BLOG by the id
-         const blog = await blogModel.findByIdAndUpdate(id, newBlogData)
+         const blog = await blogModel.findByIdAndUpdate(id, newBlogData, {new:true})
          res.status(202).json(blog)
      } catch (error) {
          console.log(error)
@@ -105,7 +108,13 @@ router.put('/:id',authMiddleware ,async (req, res) => {
 router.delete('/:id', authMiddleware,async (req, res) => {
     const id = req.params.id
 
-    try {
+   try {
+    const blogtoUpdate = await blogModel.findById(id) 
+         
+    if (blogtoUpdate.creator_id.toString()  !==  req.user.id ) {
+        return res.status(400).json({msg: 'Not Authorized ! '})
+    } 
+        
         const blog= await blogModel.findByIdAndDelete(id)
         res.status(200).json( {msg: `Blog # ${id} was deleted`})
     } catch (error) {
